backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pandas
matplotlib
